Add jit allocation functions
This jit internally uses a pooled allocation of empty jit blocks.
Each free block contains a reference to the next 
except for the last one;
that one has the special value of one more then
the highest block number.
When the jit runs out of blocks, the least called half of all jit blocks
get deallocated.